Docker fixes for Pi

+++ buildx.sh
@@ -11,12 +11,15 @@ echo "Using registry [$REGISTRY]"
 echo -e "[registry.\"$REGISTRY\"]\n\tmirrors = [\"$REGISTRY\"]\n\thttp = true\n\tinsecure = true" > ./buildx.toml
 
 # configure the buildx environment
-export DOCKER_CLI_EXPERIMENTAL=enabled
-docker run --rm --privileged docker/binfmt:a7996909642ee92942dcd6cff44b9b95f08dad64
-docker buildx rm games
-docker buildx create --name games --config ./buildx.toml
-docker buildx use games
-docker buildx inspect --bootstrap
+if [ ! "$(docker ps -a | grep buildx_buildkit_games)" ]
+then
+    export DOCKER_CLI_EXPERIMENTAL=enabled
+    docker run --rm --privileged docker/binfmt:a7996909642ee92942dcd6cff44b9b95f08dad64
+    docker buildx rm games
+    docker buildx create --name games --config ./buildx.toml
+    docker buildx use games
+    docker buildx inspect --bootstrap
+fi
 
 # build the API
 echo "Building API"
@@ -27,7 +30,7 @@ echo "Done"
 
 # build the UI
 echo "Building UI"
-cd api
+cd ui
 docker buildx build --platform linux/amd64,linux/arm/v7 --push -t $REGISTRY/games/ui:0.0.1 .
 cd ..
 echo "Done"
